[storage] Validate snapshot update disk file count
